Split out CoverageCommand with finder stuff
